requirements.txt
opencv-contrib-python
pycryptodome